pyproject.toml
[project]
name = "massband"
version = "0.1.1"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "jax>=0.5.0",
    "kinisi>=1.1.1",
    "laufband>=0.1.3",
    "lazy-loader>=0.4",
    "matplotlib>=3.10.3",
    "pint>=0.23",
    "plotly>=6.2.0",
    "rdkit2ase>=0.1.11",
    "uravu>=1.3.0",
    "znh5md>=0.4.8",
    "zntrack>=0.8.8",
]
license = "MIT"
[project.optional-dependencies]
cuda = ["jax[cuda]"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "pytest>=8.3.5",
    "tidynamics>=1.1.2",
    "zndraw>=0.5.10",
    "dvc-s3>=3.2.2",
    "nbformat>=5.10.4",
]


[tool.ruff]
line-length = 90

[tool.ruff.lint]
select = ["F", "C", "I"] #, "ANN"]
